First implementation of FluentAssertions.OneOf
Only OneOf<> is supported as Type inference does not work on constraint types, ruling out OneOfBase<>
To use type inference all the types must be contained within the parameters.
I have written the extension methods .Should() Which returns a OneOfAssertion<>
Be<TResult> is an extension method which attempts to make sure the value of the OneOf is the given case.
It returns And<TResult> because I can't generically return an item which has been wrapped in an assertion because extension methods are used to initialise them.
Plus introduced a test lib without fleshing out the tests yet.